Refactor - Application Factory Pattern
