+++ api/models/itemmodel.py
@@ -1,4 +1,4 @@
-from db import db
+from api.db import db
 
 
 class ItemModel(db.Model):
