+++ api/resources/user.py
@@ -2,7 +2,7 @@ from flask_restful import Resource
 from flask import request
 from flask_jwt_extended import create_access_token, create_refresh_token, jwt_required
 
-from models.usermodel import UserModel
+from api.models.usermodel import UserModel
 
 
 class UserRegister(Resource):
